# Faster CPU Name Retrieval Method