- Replaces the CIM method with one that uses the registry
  - Uses the keys `~MHz` and `ProcessorNameString` at `HKLM:\HARDWARE\DESCRIPTION\System\CentralProcessor\0`
    - Uses `[Microsoft.Win32.RegistryKey]::OpenRemoteBaseKey('LocalMachine', $Computername).OpenSubKey("HARDWARE\DESCRIPTION\System\CentralProcessor\0")` and `GetValue()` to get the registry keys much faster than using `Get-ItemProperty`
  - Currently doesn't round the frequency as rounding is slow

Speed :
- PowerShell 5:
  - 11.71ms → 4.99ms
    - 2.35x Faster
- PowerShell 7:
  - 12.18ms → 4.23ms
    - 2.88x Faster

Bottlenecked Speed:
- PowerShell 5:
  - 39.25ms → 20.28ms
    - 1.94x Faster
- PowerShell 7:
  - 29.43ms → 15.17ms
    - 1.94x Faster