@rashil2000 2023-01-06
It seems to be showing frequency twice:
![image](https://user-images.githubusercontent.com/46838874/211081150-7b31cc48-60b7-4b3a-ae5b-cb3a31ab8752.png)

Also, please don't use camelcase. Stick to small case.