@Carterpersall 2023-02-01
> > Uses `[Microsoft.Win32.RegistryKey]::OpenRemoteBaseKey('LocalMachine', $Computername).OpenSubKey("HARDWARE\DESCRIPTION\System\CentralProcessor\0")` and `GetValue()` to get the registry keys much faster than using `Get-ItemProperty`
> 
> @Carterpersall off topic but there are a bunch of other places where we use `Get-ItemProperty` to read the registry - do you think we should replace them too?

Might be worth checking if replacing them causes as much of a speed improvement as this does